[#89331264] Implement tests for WaldorfAndStatler
Found solutions to my problem of not being able to directly
test WaldorfAndStatler.review functionality.  As it can't be directly
called, I had to implement a dummy class around the module call.

These do seem to duplicate similar tests in playlist_spec, but since the
Playlist.play method potentially does more than call this module, it
could be argued that they are not strict duplicates.